sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "redstone-sdk"
version = "1.0.0"
description = "CLI SDK for Redstone Multi-Agent Development Platform"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Redstone", email = "dev@redstone.dev"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
redstone = "redstone_sdk.cli:app"

[project.urls]
Homepage = "https://redstone.dev"
Documentation = "https://docs.redstone.dev"
Repository = "https://github.com/redstone-dev/redstone-sdk"
